refacotor and add relation
