+++ server/api/files.py
@@ -25,6 +25,7 @@ def get_file(id):
 
 @file_bp.route("", methods=["POST"])
 def create_file():
+
     r = request.get_json(force=True)
     if not r:
         abort(400)
