--- server/schemas/__init__.py
@@ -1,14 +1,0 @@
-from server.schemas.task_schema import TaskSchema
-from server.schemas.file_schema import FileSchema
-from server.schemas.user_schema import UserSchema
-from server.schemas.task_schema import TaskSchema
-from server.schemas.visit_schema import VisitSchema
-from server.schemas.web_schema import WebSchema
-
-
-__all__ = [TaskSchema,
-           FileSchema,
-           UserSchema,
-           TaskSchema,
-           VisitSchema,
-           WebSchema]
